convert select tests to vitest
